requirements.txt
python-telegram-bot[job-queue]==21.10
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1
